requirements.txt
numpy>=1.11.3
matplotlib>=2.2.2
pillow>=6.2.0
scipy>=0.18.1
networkx>=2.5.1
scikit-learn>=0.24.1
tikzplotlib~=0.9.12
pytest

